+++ src/main/java/fr/zazac1/customrecipe/mixin/HolderReferenceInvoker.java
@@ -1,0 +1,11 @@
+package fr.zazac1.customrecipe.mixin;
+
+import net.minecraft.core.Holder;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(Holder.Reference.class)
+public interface HolderReferenceInvoker<T> {
+    @Invoker("bindValue")
+    void customrecipe$bindValue(T value);
+}